why must i always do this
