+++ index.html
@@ -1,6 +1,7 @@
 <html>
   <head>
     <meta charset="UTF-8">
+    <base href="/">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link rel="stylesheet" href="./style.css" type="text/css" />
     <title></title>
